requirements.txt
supabase
feedparser
groq
python-dotenv

